test: segregate unchallenged_exit by bad block from just being late
fixup move more tests
